Declare security role in web.xml
Avoid warning during Tomcat startup.

From the web-app.xsd documentation :

The role-name used here must either correspond to the role-name of one
of the security-role elements defined for this web application,
or be the specially	reserved role-name "*" that is a compact syntax for
indicating all roles in the web application.